+++ datanode/gateway/graphql/resolvers.go
@@ -1755,7 +1755,9 @@ func (r *myPartyResolver) AccountsConnection(ctx context.Context, party *vegapb.
 		accTy = *accType
 		if accTy != vegapb.AccountType_ACCOUNT_TYPE_GENERAL &&
 			accTy != vegapb.AccountType_ACCOUNT_TYPE_MARGIN &&
-			accTy != vegapb.AccountType_ACCOUNT_TYPE_BOND {
+			accTy != vegapb.AccountType_ACCOUNT_TYPE_BOND &&
+			accTy != vega.AccountType_ACCOUNT_TYPE_VESTED_REWARDS &&
+			accTy != vega.AccountType_ACCOUNT_TYPE_VESTING_REWARDS {
 			return nil, fmt.Errorf("invalid account type for party %v", accType)
 		}
 		accountTypes = []vegapb.AccountType{accTy}